requirements.txt
pwntools
termcolor